django_api/requirements.txt
app==0.0.1
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
distlib==0.3.8
Django==5.0.6
django-appconf==1.0.6
django-bootstrap4==24.3
django-compressor==4.4
django-cors-headers==4.3.1
django-filter==24.2
django-postgresql==0.0.3
django-tailwind==3.8.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
filelock==3.14.0
gunicorn==22.0.0
idna==3.7
packaging==24.1
pillow==10.3.0
platformdirs==4.2.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
rcssmin==1.1.1
requests==2.31.0
rjsmin==1.2.1
soupsieve==2.5
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.26.1
whitenoise==6.7.0
